zalloc needs to support large chunks of memory

+++ usr/src/boot/libsa/zalloc_mem.h
@@ -38,12 +38,12 @@
 
 struct MemPool;
 
-typedef void *(zalloc_alloc_t)(struct MemPool *, uintptr_t, intptr_t *);
+typedef void *(zalloc_alloc_t)(struct MemPool *, uintptr_t, size_t *);
 typedef void (zalloc_free_t)(void *);
 
 typedef struct MemNode {
 	struct MemNode	*mr_Next;
-	uintptr_t	mr_Bytes;
+	uint64_t	mr_Bytes;
 } MemNode;
 
 typedef struct MemPool {
@@ -53,8 +53,8 @@ typedef struct MemPool {
 	void		*mp_Base;
 	void		*mp_End;
 	MemNode		*mp_First;
-	uintptr_t	mp_Size;
-	uintptr_t	mp_Used;
+	uint64_t	mp_Size;
+	uint64_t	mp_Used;
 	struct MemPool	*mp_next;	/* Next segment */
 } MemPool;
